refactor(update): modify to update grid after search
