+++ core/templates/dynamic_grid.html
@@ -109,25 +109,30 @@
   </div>
 </div>
 
-<!-- jQuery (required for Select2) -->
-<script src="https://code.jquery.com/jquery-3.6.0.min.js"></script>
-<!-- Select2 CSS/JS for inputable dropdowns -->
-<link href="https://cdn.jsdelivr.net/npm/select2@4.1.0-rc.0/dist/css/select2.min.css" rel="stylesheet" />
-<link href="https://cdn.jsdelivr.net/npm/select2-bootstrap-5-theme@1.3.0/dist/select2-bootstrap-5-theme.min.css" rel="stylesheet" />
-<script src="https://cdn.jsdelivr.net/npm/select2@4.1.0-rc.0/dist/js/select2.min.js"></script>
-<!-- Dynamic Grid CSS/JS -->
-<link rel="stylesheet" href="{% static 'core/css/dynamic_grid.css' %}">
-<script src="{% static 'core/js/dynamic_grid.js' %}"></script>
-<!-- Bootstrap Toast for notifications -->
-<div class="position-fixed bottom-0 end-0 p-3" style="z-index: 1100">
-  <div id="grid-toast" class="toast align-items-center text-bg-success border-0" role="alert" aria-live="assertive" aria-atomic="true" data-bs-delay="3000">
-    <div class="d-flex">
-      <div class="toast-body" id="grid-toast-body">
-        <!-- Message will be injected here -->
-      </div>
-      <button type="button" class="btn-close btn-close-white me-2 m-auto" data-bs-dismiss="toast" aria-label="Close"></button>
-    </div>
-  </div>
-</div>
+            <!-- jQuery (required for Select2) -->
+            <script src="https://code.jquery.com/jquery-3.6.0.min.js"></script>
+            <!-- Select2 CSS/JS for inputable dropdowns -->
+            <link href="https://cdn.jsdelivr.net/npm/select2@4.1.0-rc.0/dist/css/select2.min.css" rel="stylesheet" />
+            <link href="https://cdn.jsdelivr.net/npm/select2-bootstrap-5-theme@1.3.0/dist/select2-bootstrap-5-theme.min.css" rel="stylesheet" />
+            <script src="https://cdn.jsdelivr.net/npm/select2@4.1.0-rc.0/dist/js/select2.min.js"></script>
+            <!-- Bootstrap JS -->
+            <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/js/bootstrap.bundle.min.js"></script>
+            <!-- Dynamic Grid CSS -->
+            <link rel="stylesheet" href="{% static 'core/css/dynamic_grid.css' %}">
+            <!-- Modal logic (non-module, attaches to window) -->
+            <script src="{% static 'core/js/dynamic_grid/modal.js' %}"></script>
+            <!-- Main dynamic grid logic as a module -->
+            <script type="module" src="{% static 'core/js/dynamic_grid.js' %}"></script>
+            <!-- Bootstrap Toast for notifications -->
+            <div class="position-fixed bottom-0 end-0 p-3" style="z-index: 1100">
+              <div id="grid-toast" class="toast align-items-center text-bg-success border-0" role="alert" aria-live="assertive" aria-atomic="true" data-bs-delay="3000">
+                <div class="d-flex">
+                  <div class="toast-body" id="grid-toast-body">
+                    <!-- Message will be injected here -->
+                  </div>
+                  <button type="button" class="btn-close btn-close-white me-2 m-auto" data-bs-dismiss="toast" aria-label="Close"></button>
+                </div>
+              </div>
+            </div>
 {% endif %}
 {% endblock %}
