+++ core/views.py
@@ -335,4 +335,12 @@ def api_search(request, table_name):
     count_sql = f"SELECT COUNT(*) FROM {table_name} {where_sql}"
     cursor.execute(count_sql, params)
     total_count = cursor.fetchone()[0]
+    return render(request, 'dynamic_grid.html', {
+        'columns': columns,  # list of (label, field_name) to display
+        'data': data,        # list of dicts, field_name -> value
+        'form_name': form_name,
+        'table_name': table_name,
+        'menu_tree': menu_tree,
+        'total_count': total_count,
+    })
     return JsonResponse({'columns': columns, 'data': data, 'total_count': total_count})
